backend/requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2.20
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
django-timezone-field==7.1
django-unfold==0.58.0
django_celery_results==2.6.0
djangochannelsrestframework==1.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
msgpack==1.1.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
